renaming complete

+++ app/services/user_service/user_likes_music_canceler.rb
@@ -8,6 +8,9 @@ module UserService
         def call
             begin
                 destroy_music = UserLikesMusic.find_by(user_id: @user_id, music_id: @music_id)
+                if destroy_music.nil?
+                    return {success: false, message: "User is not liking this music"}
+                end
                 destroy_music.destroy
                 return {success: true}
             rescue => e
